fix(lint): clippy kapısını yeşile çek — maskelenmiş 12 uyarı
`cargo clippy --workspace` oturum boyunca **ilk crate'te hata verip
duruyordu** (`doc_lazy_continuation`, moneywar-domain). Ben bu süre boyunca
crate-filtreli grep'lerle bakıp "clippy temiz" diye rapor ettim; filtre
eşleşmediği için boş çıktıyı temizlik sanmışım. Gerçekte gerideki tüm crate'ler
hiç denetlenmiyordu.

Kapı açılınca çıkanlar düzeltildi. İkisi gerçek bulgu:

- `market.rs`: Walras faktöründe `if imbalance > 0 { A } else { A }` — iki dal
  birebir aynıydı. Muhtemelen asimetrik ayar (fiyat yukarı hızlı, aşağı yavaş)
  düşünülüp yazılmamış. `imbalance` işaretli olduğu için tek ifade iki yönü de
  doğru işliyor; davranış değişmeden sadeleştirildi, niyet nota alındı.
- `main.rs`: şehir profili çarpanında üç dal da 2 dönüyordu (v8.5'te `demand`
  1→2 çıkarılınca fallback ile eşitlenmiş). Davranış korunarak tek dala indi.

Kalanlar mekanik: `clamp` kalıpları, gereksiz closure, `map_or` sadeleştirmesi,
karmaşık dönüş tiplerine `type` alias (`EventBucket`, `ChurnBuckets`), doküman
liste girintileri.

Testler: 26 test suite, hepsi geçiyor. Denge değişmedi (makas 3.3×).

+++ crates/moneywar-cli/src/chatter.rs
@@ -142,7 +142,7 @@ fn shock_for(state: &GameState, tick: Tick, event: &LogEvent) -> Option<ChatterE
     else {
         return None;
     };
-    let speaker_id = pick_npc_speaker(state, tick, u64::from(event_id.value()))?;
+    let speaker_id = pick_npc_speaker(state, tick, event_id.value())?;
     let speaker = state.players.get(&speaker_id)?;
     let personality = speaker.personality.unwrap_or(Personality::TrendFollower);
     let kind = ChatterKind::ShockEvent;
